+++ src/domain/Datatable.js
@@ -140,11 +140,12 @@ export default class Datatable extends BidEntity {
     /**
      * Exports datatable to core structure.
      * 
+     * @param {boolean} [alwaysIncludeConfig=false] Will include config object in export regardless of changed status.
      * @returns {object}
      */
-    exportData() {
+    exportData(alwaysIncludeConfig=false) {
         let data = _.cloneDeep(this._data);
-        if (_.isEqual(data.config, this._original.config)) delete data.config;
+        if (!alwaysIncludeConfig && _.isEqual(data.config, this._original.config)) delete data.config;
 
         return data;
     }